Created generators for all the data for the products
Created random generators to generate all the data for the products to be able to create more products more easily. Thus deleted the created array of product data that was used previously. Needed to add a context for the various elements to access the data. Also deleted the .map that renders products in sequence so now only one product is executed