Reworked how products are generated
Reworked how products are generated to make the shopping cart solution easier to solve. Created json string instead of an array. In order to loop out products, however, a for loop was created which added the products to an array but because the array was then converted back to a json string in the end.