Created shopping cart functionality so you can add products
Redid the cart data from scratch and used the react-use-cart setup. Add addItem to perfumesite and when a user clicks add to cart, addItem sends item (the product data). Added that the writeProducts element sends the data through e.g. name={item.name} and not through data{products} (so the whole array). Also added styling for the shopping cart page.