Created localstorage functionality to get the same products all the time
Created functionality for the same products to be generated every time so that measurements are correct. Created it through localstorage by retuning the data from the json string. Create a function that saves the json string in localstorages through setItem. Then the json string is retrieved through getItem and then an if clause is used which says that if getItem does not have products, the function that saves the json string must be executed. Then the variable product localstorages is given with the json string.