Fixed the men's perfume page
Have created a json string for the data for Mr. Perfumes so now everything works as it should. Did the same way as for the women's perfumes with new functions and variable names. Deleted some functions for the component consisting of functions used previously for the men's perfumes. These have been put in the mr component which generates a json string.