Created the final functionality for the products to be added to the cart.
Created the final functionality for the products to be added to the cart. However, this is only created for the women's perfumes at the moment. Thus deleted use-cart-react as it is a third-party library and could have affected my measurements, which makes me choose to write the code manually. Created a function called addToCart that adds the clicked product to setCartProduct.