Added so that the men's perfumes can also be added to the shopping cart
Added so that the men's perfumes can also be added to the shopping cart + that the total price changes. Also changed which lady the perfumes had with id to get uniqueness which sometimes didn't happen before. Changed so that all names in the json strings, regardless of whether it is female or male, have the same name to be able to print the data in one way in the shopping cart.