Add to see the total number of a product added
Add to see the total number of a product added to make both projects look the same. What was done was that I added an if statement to the addToCart function to check if the product is already in the cart through -1 index. If it is already in the shopping cart, it updates the quantity with +1, if not, it adds the product to the shopping cart and sets the product quantity to 1.